+++ EarthForces/EarthGunnery.h
@@ -5,7 +5,7 @@ class EarthGunnery:public unit
 {
 	
 public:
-	EarthGunnery(int H, int P, int AC, int T);
+	EarthGunnery(double H, int P, int AC, int T);
 	bool attack(Game* GPtr);
 };
 #endif
